modified main.rs to skip theme that exists

+++ crates/theme_importer/src/main.rs
@@ -318,6 +318,12 @@ fn main() -> Result<()> {
             .replace(")", "")
             .to_case(Case::Snake);
 
+        let output_file_path = themes_output_path.join(format!("{theme_family_slug}.rs"));
+        if (output_file_path.exists()) {
+            log::info!("Skipping file {theme_family_slug}.rs, it already exists. In order to update a theme, please delete it from the output path {theme_output_path}.");
+            continue;
+        }
+
         let mut output_file =
             File::create(themes_output_path.join(format!("{theme_family_slug}.rs")))?;
         log::info!(
